requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
openai>=1.3.7
gradio>=4.7.1
python-dotenv>=1.0.0
plotly>=5.17.0
matplotlib>=3.8.2
scikit-learn>=1.3.2
numpy>=1.24.3
sentence-transformers>=2.2.2
